requirements.txt
requests
Flask
beautifulsoup4
python-dotenv
telethon